+++ mitm/core.py
@@ -2,10 +2,15 @@
 Core components of the MITM framework.
 """
 
+from __future__ import annotations
+
 import asyncio
+from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from enum import Enum
-from typing import Any, Optional
+from typing import Any, List, Optional, Tuple
+
+from .crypto import CertificateAuthority
 
 
 @dataclass
@@ -97,7 +102,7 @@ class Connection:
 
     A connection is a pair of `Host` objects that the `mitm` relays data between. When
     a connection is created the server host is not resolved until the data is
-    intercepted and the destination server is figured out.
+    intercepted and the protocol and destination server is figured out.
 
     Note:
         See more on `dataclasses <https://docs.python.org/3/library/dataclasses.html>`_.
@@ -105,6 +110,7 @@ class Connection:
     Args:
         client: The client host.
         server: The server host.
+        protocol: The protocol of the connection.
 
     Example:
         .. code-block:: python
@@ -112,14 +118,15 @@ class Connection:
             client = Host(...)
             server = Host(...)
 
-            connection = Connection(client, server)
+            connection = Connection(client, server, Protocol.HTTP)
     """
 
     client: Host
     server: Host
+    protocol: Optional[Protocol] = None
 
     def __repr__(self):
-        return f"<Connection client={self.client} server={self.server}>"
+        return f"<Connection client={self.client} server={self.server}, protocol={self.protocol}>"
 
 
 class Flow(Enum):
@@ -136,3 +143,196 @@ class Flow(Enum):
 
     CLIENT_TO_SERVER = 0
     SERVER_TO_CLIENT = 1
+
+
+class Middleware(ABC):
+    """
+    Event-driven hook extension for the `mitm`.
+    """
+
+    @abstractmethod
+    async def mitm_started(self, host: str, port: int):
+        """
+        Called when the mitm has started.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    async def client_connected(self, connection: Connection):
+        """
+        Called when the connection is established with the client.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    async def server_connected(self, connection: Connection):
+        """
+        Called when the connection is established with the server.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    async def client_data(self, connection: Connection, data: bytes) -> bytes:
+        """
+        Called when data is received from the client.
+
+        Note:
+            Modifying the request will only modify the request sent to the destination
+            server, and not the request mitm interprets. In other words, modifying the
+            'Host' headers will not change the destination server.
+
+            Raw TLS/SSL handshake is not sent through this method.
+
+        Args:
+            request: The request received from the client.
+
+        Returns:
+            The request to send to the server.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    async def server_data(self, connection: Connection, data: bytes) -> bytes:
+        """
+        Called when data is received from the server.
+
+        Args:
+            response: The response received from the server.
+
+        Returns:
+            The response to send back to the client.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    async def client_disconnected(self, connection: Connection):
+        """
+        Called when the client disconnects.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    async def server_disconnected(self, connection: Connection):
+        """
+        Called when the server disconnects.
+
+        Note:
+            By the time this method is called, the server will have already successfully
+            disconnected.
+        """
+        raise NotImplementedError
+
+    def __repr__(self):
+        return f"<Middleware: {self.__class__.__name__}>"
+
+
+class InvalidProtocol(Exception):
+    """
+    Exception raised when the protocol did not work.
+
+    This is the only error that `mitm.MITM` will catch. Throwing this error will
+    continue the search for a valid protocol.
+    """
+
+
+class Protocol(ABC):
+    """
+    An abstract class for a custom protocol implementation.
+
+    The `bytes_needed` is used to determine the minimum number of bytes needed to be
+    read from the connection to identify all of the protocols. This is done by getting
+    the `max()` of the `bytes_needed` of all the protocols, and reading that many
+    bytes from the connection.
+
+    Args:
+        bytes_needed: The minimum number of bytes needed to identify the protocol.
+
+    Example:
+
+        Template for a protocol implementation:
+
+        .. code-block:: python
+
+            from mitm import Protocol, Connection
+
+            class MyProtocol(Protocol):
+                bytes_needed = 4
+
+                @classmethod
+                async def connect(cls: Protocol, connection: Connection, data: bytes) -> bool:
+                    # Do something with the data.
+    """
+
+    def __init__(
+        self,
+        bytes_needed: int = 8192,
+        buffer_size: int = 8192,
+        timeout: int = 15,
+        keep_alive: bool = True,
+        certificate_authority: CertificateAuthority = CertificateAuthority(),
+        middlewares: List[Middleware] = [],
+    ):
+        self.bytes_needed = bytes_needed
+        self.buffer_size = buffer_size
+        self.timeout = timeout
+        self.keep_alive = keep_alive
+        self.certificate_authority = certificate_authority
+        self.middlewares = middlewares
+
+    @abstractmethod
+    async def resolve(self, connection: Connection, data: bytes) -> Optional[Tuple[str, int, bool]]:
+        """
+        Resolves the destination of the connection.
+
+        Args:
+            connection: Connection object containing a client host.
+            data: The initial incoming data from the client.
+
+        Returns:
+            A tuple containing the host, port, and bool if the connection is encrypted.
+
+        Raises:
+            InvalidProtocol: If the connection failed.
+
+        Note:
+            This methods needs to be implemented by subclasses.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    async def connect(self, connection: Connection, host: str, port: int, tls: bool, data: bytes):
+        """
+        Attempts to connect to destination server using the given data. Returns `True`
+        if the connection was successful, raises `InvalidProtocol` if the connection
+        failed.
+
+        Args:
+            connection: Connection object containing a client host.
+            data: The initial incoming data from the client.
+
+        Returns:
+            Whether the connection was successful.
+
+        Raises:
+            InvalidProtocol: If the connection failed.
+
+        Note:
+            This methods needs to be implemented by subclasses.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    async def handle(self, connection: Connection):
+        """
+        Handles the connection between a client and a server.
+
+        Args:
+            connection: Client/server connection to relay.
+
+        Note:
+            This methods needs to be implemented by subclasses.
+        """
+        raise NotImplementedError
+
+    def __repr__(self):
+        return f"<Protocol: {self.__class__.__name__}>"
